refactor(acceptance): extract scenario 25 helpers into sibling file
Split the file-local helper group out of
scripts/acceptance/scenarios/25-plot-roundtrip.ts into a new sibling
25-plot-roundtrip.helpers.ts, mirroring scenarios 20/28. The scenario
body, header comment, and run.ts import contract are unchanged.

warren-8a36, pl-91f7 step 1
